telegram-bot/requirements.txt
python-telegram-bot>=20.0
qrcode[pil]>=7.4.2
python-dotenv>=1.0.0
